split capability mapping into dedicated microcrate
