+++ crates/perl-lsp-capability-map/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "perl-lsp-capability-map"
+version = "0.10.0"
+edition = { workspace = true }
+rust-version = { workspace = true }
+authors = { workspace = true }
+license = "MIT OR Apache-2.0"
+repository = { workspace = true }
+homepage = { workspace = true }
+description = "SRP microcrate for translating between LSP ServerCapabilities and feature IDs."
+documentation = "https://docs.rs/perl-lsp-capability-map"
+readme = "README.md"
+keywords = ["perl", "lsp", "capabilities", "features", "mapping"]
+categories = ["development-tools"]
+include = ["src/**", "Cargo.toml", "README.md"]
+
+[dependencies]
+lsp-types = "0.97.0"
+perl-lsp-feature-ids = { workspace = true }
+
+[lints]
+workspace = true
