+++ crates/perl-lsp-feature-contracts/Cargo.toml
@@ -19,6 +19,7 @@ serde = { workspace = true }
 lsp-types = "0.97.0"
 perl-feature-catalog = { workspace = true }
 perl-lsp-feature-ids = { workspace = true }
+perl-lsp-capability-map = { workspace = true }
 
 [build-dependencies]
 perl-feature-catalog = { workspace = true }
