+++ crates/perl-lsp-feature-contracts/src/lib.rs
@@ -5,7 +5,7 @@
 //! between `perl-lsp-feature-ids` (raw identifiers) and
 //! `perl-lsp-feature-policy` (runtime capability selection).
 
-use perl_lsp_feature_ids::*;
+pub use perl_lsp_capability_map::{caps_from_feature_ids, feature_ids_from_caps};
 use serde::Serialize;
 
 /// Canonical metadata for profile aliases and normalization behavior.
@@ -117,8 +117,6 @@ pub mod catalog {
     include!(concat!(env!("OUT_DIR"), "/feature_contracts.rs"));
 }
 
-use lsp_types::ServerCapabilities;
-
 /// Human-readable BDD-oriented feature row for automation and reporting.
 #[derive(Debug, Clone, Serialize)]
 pub struct BddFeatureRow {
@@ -187,338 +185,4 @@ pub fn compliance_percent_for_grid() -> f32 {
     }
     let advertised = advertised_trackable_feature_count_for_grid();
     (advertised as f64 / trackable as f64 * 100.0).round() as f32
-}
-
-/// Extract feature IDs from LSP `ServerCapabilities`.
-pub fn feature_ids_from_caps(c: &ServerCapabilities) -> Vec<&'static str> {
-    let mut v = Vec::new();
-
-    // Text Document Features
-    if c.completion_provider.is_some() {
-        v.push(LSP_COMPLETION);
-    }
-    if c.hover_provider.is_some() {
-        v.push(LSP_HOVER);
-    }
-    if c.signature_help_provider.is_some() {
-        v.push(LSP_SIGNATURE_HELP);
-    }
-    if c.definition_provider.is_some() {
-        v.push(LSP_DEFINITION);
-    }
-    if c.declaration_provider.is_some() {
-        v.push(LSP_DECLARATION);
-    }
-    if c.notebook_document_sync.is_some() {
-        v.push(LSP_NOTEBOOK_DOCUMENT_SYNC);
-    }
-    if c.type_definition_provider.is_some() {
-        v.push(LSP_TYPE_DEFINITION);
-    }
-    if c.implementation_provider.is_some() {
-        v.push(LSP_IMPLEMENTATION);
-    }
-    if c.references_provider.is_some() {
-        v.push(LSP_REFERENCES);
-    }
-    if c.document_highlight_provider.is_some() {
-        v.push(LSP_DOCUMENT_HIGHLIGHT);
-    }
-    if c.document_symbol_provider.is_some() {
-        v.push(LSP_DOCUMENT_SYMBOL);
-    }
-    if c.code_action_provider.is_some() {
-        v.push(LSP_CODE_ACTION);
-    }
-    if c.code_lens_provider.is_some() {
-        v.push(LSP_CODE_LENS);
-    }
-    if c.document_link_provider.is_some() {
-        v.push(LSP_DOCUMENT_LINK);
-    }
-    if c.color_provider.is_some() {
-        v.push(LSP_DOCUMENT_COLOR);
-    }
-    if c.document_formatting_provider.is_some() {
-        v.push(LSP_FORMATTING);
-    }
-    if c.document_range_formatting_provider.is_some() {
-        v.push(LSP_RANGE_FORMATTING);
-    }
-    if c.document_on_type_formatting_provider.is_some() {
-        v.push(LSP_ON_TYPE_FORMATTING);
-    }
-    if c.rename_provider.is_some() {
-        v.push(LSP_RENAME);
-    }
-    if c.folding_range_provider.is_some() {
-        v.push(LSP_FOLDING_RANGE);
-    }
-    if c.selection_range_provider.is_some() {
-        v.push(LSP_SELECTION_RANGE);
-    }
-    if c.linked_editing_range_provider.is_some() {
-        v.push(LSP_LINKED_EDITING_RANGE);
-    }
-    if c.call_hierarchy_provider.is_some() {
-        v.push(LSP_CALL_HIERARCHY);
-    }
-    if c.semantic_tokens_provider.is_some() {
-        v.push(LSP_SEMANTIC_TOKENS);
-    }
-    if c.moniker_provider.is_some() {
-        v.push(LSP_MONIKER);
-    }
-    // Note: type_hierarchy_provider doesn't exist in lsp-types 0.97
-    // This would be added in newer versions of lsp-types
-    if c.inline_value_provider.is_some() {
-        v.push(LSP_INLINE_VALUE);
-    }
-    if c.inlay_hint_provider.is_some() {
-        v.push(LSP_INLAY_HINT);
-    }
-    if c.diagnostic_provider.is_some() {
-        v.push(LSP_PULL_DIAGNOSTICS);
-    }
-
-    // Workspace Features
-    if c.workspace_symbol_provider.is_some() {
-        v.push(LSP_WORKSPACE_SYMBOL);
-    }
-    if c.execute_command_provider.is_some() {
-        v.push(LSP_EXECUTE_COMMAND);
-    }
-
-    // Note: Some features like workspace edit, file operations etc. are in workspace capabilities
-    // which are separate from ServerCapabilities
-
-    v.sort();
-    v.dedup();
-    v
-}
-
-/// Build LSP `ServerCapabilities` from feature IDs.
-pub fn caps_from_feature_ids(features: &[&str]) -> ServerCapabilities {
-    use lsp_types::*;
-
-    let mut caps = ServerCapabilities::default();
-
-    for &feature in features {
-        match feature {
-            LSP_COMPLETION => {
-                caps.completion_provider = Some(CompletionOptions {
-                    trigger_characters: Some(vec![
-                        "$".to_string(),
-                        "@".to_string(),
-                        "%".to_string(),
-                        ">".to_string(),
-                        ":".to_string(),
-                    ]),
-                    ..Default::default()
-                });
-            }
-            LSP_HOVER => {
-                caps.hover_provider = Some(HoverProviderCapability::Simple(true));
-            }
-            LSP_SIGNATURE_HELP => {
-                caps.signature_help_provider = Some(SignatureHelpOptions {
-                    trigger_characters: Some(vec!["(".to_string(), ",".to_string()]),
-                    ..Default::default()
-                });
-            }
-            LSP_DEFINITION => {
-                caps.definition_provider = Some(OneOf::Left(true));
-            }
-            LSP_DECLARATION => {
-                caps.declaration_provider = Some(DeclarationCapability::Simple(true));
-            }
-            LSP_NOTEBOOK_DOCUMENT_SYNC => {
-                caps.notebook_document_sync = Some(OneOf::Left(NotebookDocumentSyncOptions {
-                    notebook_selector: vec![NotebookSelector::ByNotebook {
-                        notebook: Notebook::String("jupyter-notebook".to_string()),
-                        cells: Some(vec![NotebookCellSelector { language: "perl".to_string() }]),
-                    }],
-                    save: Some(true),
-                }));
-            }
-            LSP_TYPE_DEFINITION => {
-                caps.type_definition_provider =
-                    Some(TypeDefinitionProviderCapability::Simple(true));
-            }
-            LSP_IMPLEMENTATION => {
-                caps.implementation_provider = Some(ImplementationProviderCapability::Simple(true));
-            }
-            LSP_REFERENCES => {
-                caps.references_provider = Some(OneOf::Left(true));
-            }
-            LSP_DOCUMENT_SYMBOL => {
-                caps.document_symbol_provider = Some(OneOf::Left(true));
-            }
-            LSP_CODE_ACTION => {
-                caps.code_action_provider = Some(CodeActionProviderCapability::Simple(true));
-            }
-            LSP_FORMATTING => {
-                caps.document_formatting_provider = Some(OneOf::Left(true));
-            }
-            LSP_RANGE_FORMATTING => {
-                caps.document_range_formatting_provider = Some(OneOf::Left(true));
-            }
-            LSP_RENAME => {
-                caps.rename_provider = Some(OneOf::Left(true));
-            }
-            LSP_FOLDING_RANGE => {
-                caps.folding_range_provider = Some(FoldingRangeProviderCapability::Simple(true));
-            }
-            LSP_SEMANTIC_TOKENS => {
-                caps.semantic_tokens_provider =
-                    Some(SemanticTokensServerCapabilities::SemanticTokensOptions(
-                        SemanticTokensOptions {
-                            legend: SemanticTokensLegend {
-                                token_types: vec![
-                                    SemanticTokenType::NAMESPACE,
-                                    SemanticTokenType::TYPE,
-                                    SemanticTokenType::CLASS,
-                                    SemanticTokenType::ENUM,
-                                    SemanticTokenType::INTERFACE,
-                                    SemanticTokenType::STRUCT,
-                                    SemanticTokenType::TYPE_PARAMETER,
-                                    SemanticTokenType::PARAMETER,
-                                    SemanticTokenType::VARIABLE,
-                                    SemanticTokenType::PROPERTY,
-                                    SemanticTokenType::ENUM_MEMBER,
-                                    SemanticTokenType::EVENT,
-                                    SemanticTokenType::FUNCTION,
-                                    SemanticTokenType::METHOD,
-                                    SemanticTokenType::MACRO,
-                                    SemanticTokenType::KEYWORD,
-                                    SemanticTokenType::MODIFIER,
-                                    SemanticTokenType::COMMENT,
-                                    SemanticTokenType::STRING,
-                                    SemanticTokenType::NUMBER,
-                                    SemanticTokenType::REGEXP,
-                                    SemanticTokenType::OPERATOR,
-                                ],
-                                token_modifiers: vec![
-                                    SemanticTokenModifier::DECLARATION,
-                                    SemanticTokenModifier::DEFINITION,
-                                    SemanticTokenModifier::READONLY,
-                                    SemanticTokenModifier::STATIC,
-                                    SemanticTokenModifier::DEPRECATED,
-                                    SemanticTokenModifier::ABSTRACT,
-                                    SemanticTokenModifier::ASYNC,
-                                    SemanticTokenModifier::MODIFICATION,
-                                    SemanticTokenModifier::DOCUMENTATION,
-                                    SemanticTokenModifier::DEFAULT_LIBRARY,
-                                ],
-                            },
-                            full: Some(SemanticTokensFullOptions::Bool(true)),
-                            range: Some(true),
-                            ..Default::default()
-                        },
-                    ));
-            }
-            LSP_DOCUMENT_HIGHLIGHT => {
-                caps.document_highlight_provider = Some(OneOf::Left(true));
-            }
-            LSP_CODE_LENS => {
-                caps.code_lens_provider = Some(CodeLensOptions { resolve_provider: Some(true) });
-            }
-            LSP_DOCUMENT_LINK => {
-                caps.document_link_provider = Some(DocumentLinkOptions {
-                    resolve_provider: Some(true),
-                    work_done_progress_options: WorkDoneProgressOptions::default(),
-                });
-            }
-            LSP_DOCUMENT_COLOR | LSP_COLOR => {
-                caps.color_provider = Some(ColorProviderCapability::Simple(true));
-            }
-            LSP_ON_TYPE_FORMATTING => {
-                caps.document_on_type_formatting_provider = Some(DocumentOnTypeFormattingOptions {
-                    first_trigger_character: ";".to_string(),
-                    more_trigger_character: Some(vec!["}".to_string()]),
-                });
-            }
-            LSP_SELECTION_RANGE => {
-                caps.selection_range_provider =
-                    Some(SelectionRangeProviderCapability::Simple(true));
-            }
-            LSP_LINKED_EDITING_RANGE => {
-                caps.linked_editing_range_provider =
-                    Some(LinkedEditingRangeServerCapabilities::Simple(true));
-            }
-            LSP_CALL_HIERARCHY => {
-                caps.call_hierarchy_provider = Some(CallHierarchyServerCapability::Simple(true));
-            }
-            LSP_MONIKER => {
-                caps.moniker_provider =
-                    Some(OneOf::Right(MonikerServerCapabilities::Options(MonikerOptions {
-                        work_done_progress_options: WorkDoneProgressOptions::default(),
-                    })));
-            }
-            LSP_INLINE_VALUE => {
-                caps.inline_value_provider = Some(OneOf::Right(
-                    InlineValueServerCapabilities::Options(InlineValueOptions::default()),
-                ));
-            }
-            LSP_INLAY_HINT => {
-                caps.inlay_hint_provider =
-                    Some(OneOf::Right(InlayHintServerCapabilities::Options(InlayHintOptions {
-                        resolve_provider: Some(true),
-                        ..Default::default()
-                    })));
-            }
-            LSP_PULL_DIAGNOSTICS => {
-                caps.diagnostic_provider =
-                    Some(DiagnosticServerCapabilities::Options(DiagnosticOptions {
-                        identifier: Some("perl-lsp".to_string()),
-                        inter_file_dependencies: true,
-                        workspace_diagnostics: true,
-                        ..Default::default()
-                    }));
-            }
-            LSP_WORKSPACE_SYMBOL => {
-                caps.workspace_symbol_provider = Some(OneOf::Left(true));
-            }
-            LSP_EXECUTE_COMMAND => {
-                caps.execute_command_provider = Some(ExecuteCommandOptions {
-                    commands: vec!["perl.runCritic".to_string()],
-                    ..Default::default()
-                });
-            }
-            _ => {
-                // Unknown feature - ignore.
-            }
-        }
-    }
-
-    caps
-}
-
-#[cfg(test)]
-mod tests {
-    use lsp_types::{ColorProviderCapability, ServerCapabilities};
-
-    use super::{LSP_COLOR, LSP_DOCUMENT_COLOR, caps_from_feature_ids, feature_ids_from_caps};
-
-    #[test]
-    fn feature_ids_from_caps_reports_catalog_color_id() {
-        let caps = ServerCapabilities {
-            color_provider: Some(ColorProviderCapability::Simple(true)),
-            ..Default::default()
-        };
-
-        assert_eq!(feature_ids_from_caps(&caps), vec![LSP_DOCUMENT_COLOR]);
-    }
-
-    #[test]
-    fn caps_from_feature_ids_accepts_legacy_color_alias() {
-        let caps = caps_from_feature_ids(&[LSP_COLOR]);
-        assert!(caps.color_provider.is_some());
-    }
-
-    #[test]
-    fn caps_from_feature_ids_accepts_canonical_color_id() {
-        let caps = caps_from_feature_ids(&[LSP_DOCUMENT_COLOR]);
-        assert!(caps.color_provider.is_some());
-    }
 }
